# swagger.yaml
openapi: 3.0.0
info:
    title: Sample API
    description: Example API using Swagger/OpenAPI 3.0
    version: 1.0.0
servers:
    - url: http://localhost:3000/api/v1
      description: Development server
paths:
...
    /posts:
        get:
            summary: Get all posts
            tags:
                - Posts
            responses:
                "200":
                    description: A list of posts
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/Post"

        post:
            summary: Create a new post
            tags:
                - Posts
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                user_uuid:
                                    type: string
                                    example: f1a377c8-3f86-4eeb-afb6-abb236f7c47b
                                title:
                                    type: string
                                    example: Sample Post
                                images:
                                    type: array
                                    items:
                                        type: string
                                        format: binary
                            required:
                                - user_uuid
                                - title
                                - images
            responses:
                "201":
                    description: Post created successfully
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        type: integer
                                        example: 1
                                    message:
                                        type: string
                "400":
                    description: Invalid input
    /posts/{postId}:
        parameters:
            - in: path
              name: postId
              required: true
              schema:
                  type: integer
              description: ID of the post to retrieve
        get:
            summary: Get a post by ID
            tags:
                - Posts
            responses:
                "200":
                    description: Post details
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Post"
                "404":
                    description: Post not found

        put:
            summary: Update a post by ID
            tags:
                - Posts
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/PostUpdate"
            responses:
                "200":
                    description: Post updated successfully
                "400":
                    description: Invalid input
                "404":
                    description: Post not found

        delete:
            summary: Delete a post by ID
            tags:
                - Posts
            responses:
                "204":
                    description: Post deleted successfully
                "404":
                    description: Post not found
components:
    schemas:
...
        Post:
            type: object
            properties:
                id:
                    type: integer
                    example: 1
                user_uuid:
                    type: string
                    example: f1a377c8-3f86-4eeb-afb6-abb236f7c47b
                title:
                    type: string
                    example: Sample Post
...
        PostUpdate:
            type: object
            properties:
                title:
                    type: string
                    example: Updated Post Title
            required:
                - title
...